Adds a note for private container credentials change and rotation
A change of the credentials used to push a docker based application
requires a re-push of the application with the new credentials.
Ohterwise CF wan't be able to start this application because during start access
to the container registry is required. A start could be triggered
because of different reasons (an appliation crash or application evacuation, etc.).